# Backend/modules/auth/dto/response.py
from pydantic import BaseModel, ConfigDict
from typing import Optional

class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    expires_in: int  # segundos

class AdminResponse(BaseModel):
    id: int
    username: str
    email: str
    is_active: bool

    model_config = ConfigDict(from_attributes=True)

class MessageResponse(BaseModel):
    message: str